Disable network sync of log_IncludeTime cvar
It seems that Crytek forgot to add log_IncludeTime to Scripts/Network/cvars.txt file to prevent server from forcing value of that cvar to all connected clients. This is a big problem because all clients still use original broken implementation of the log, so if log_IncludeTime is not zero on client, some log messages will be corrupted. Thanks to Aleks for reporting this bug.